requirements.txt
numpy==1.23.0
opencv-python==4.5.5.64
Pillow==9.1.1
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyQt5-stubs==5.15.6.0
python-dotenv==0.19.2
toml==0.10.2
